+++ Tests/Common/Data/ClassicRangeConsolidatorTests.cs
@@ -22,7 +22,7 @@ namespace QuantConnect.Tests.Common.Data
 {
     public class ClassicRangeConsolidatorTests : RangeConsolidatorTests
     {
-        protected override RangeConsolidator CreateConsolidator(decimal range)
+        protected override RangeConsolidator CreateConsolidator(int range)
         {
             return new ClassicRangeConsolidator(range, x => x.Value, x => 10m);
         }
